Pre: move ValueTypeSet to core.
