+++ query-algebrizer/Cargo.toml
@@ -4,7 +4,6 @@ version = "0.0.1"
 workspace = ".."
 
 [dependencies]
-enum-set = { git = "https://github.com/rnewman/enum-set" }
 error-chain = { git = "https://github.com/rnewman/error-chain", branch = "rnewman/sync" }
 
 [dependencies.mentat_core]
