+++ query-algebrizer/src/lib.rs
@@ -8,8 +8,6 @@
 // CONDITIONS OF ANY KIND, either express or implied. See the License for the
 // specific language governing permissions and limitations under the License.
 
-extern crate enum_set;
-
 #[macro_use]
 extern crate error_chain;
 
@@ -58,7 +56,6 @@ pub use clauses::{
 
 pub use types::{
     EmptyBecause,
-    ValueTypeSet,
 };
 
 #[derive(Debug)]
