+++ query-translator/src/translate.rs
@@ -10,6 +10,7 @@
 
 use mentat_core::{
     SQLValueType,
+    SQLValueTypeSet,
     TypedValue,
     ValueType,
 };
